various code errors
incorrect phpdocs, unused variables, functions defined in the wrong class, abstract class methods, etc
