+++ customfield/field/text/classes/data.php
@@ -22,8 +22,6 @@
 
 namespace customfield_text;
 
-use core\persistent;
-
 defined('MOODLE_INTERNAL') || die;
 
 /**
@@ -36,7 +34,7 @@ class data extends \core_customfield\data {
     /**
      * Add fields for editing a text profile field.
      *
-     * @param moodleform $mform
+     * @param \MoodleQuickForm $mform
      * @throws \coding_exception
      */
     public function edit_field_add(\MoodleQuickForm $mform) {
